requirements.txt
python==3.11
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.25
passlib==1.7.4
pydantic==1.9.0
python-jose==3.3.0
python-dotenv==0.19.2
pytest==7.3.2
pytest-asyncio==0.17.0
pytest-cov==3.1.0
unittest2==1.1.0